simplify: the pass's four lenses, applied
Reuse, simplification, efficiency and altitude, run over the branch as
four independent reviews. What they found and what was done with it:

- the step file re-spelled the world's own `clickWithin`, and located
  the picker's button four ways after declaring one spelling for its
  panel and its box. One helper each now, and the click goes through the
  world.
- `{file} holds the node {id} dated {date}` was a second home for a disk
  assertion whose twin (`…with no date`) lives in `editing_steps.ts`, and
  it read the file directly where every sibling goes through `recordsIn`
  — which is what makes a file that does not exist yet read as "nothing
  there yet" instead of an ENOENT out of a helper. Moved beside its twin,
  through `recordsIn`, with the exactness argument moved with it.
- `writes.ts` re-exported `Said` "so a panel reads it from the module it
  is answered by", and the new panel then imported it from `edit/undoing.ts`
  directly. One type, one import path: the re-export is gone and the menu
  reads it where it is declared.
- three `datePick` tests were the same assertion with different literals
  — the function cannot tell setting from changing, or one id from
  another, because neither is an argument. One test, with what the other
  two were claiming in its comment.
- the picker's line about a refused write had no `aria-live`, where the
  `•••` menu's has one. A refusal interrupts a screen reader and a remark
  does not; that pair is now kept in both places.
- `pressOf`'s dead button was argued as `menu/verbs.ts`' "already so"
  rule, and that analogy does not hold: the menu's version mirrors an ops
  refusal, and `set_date` has none — an agent may re-send the date a node
  already carries. The rule it actually is is the EDITOR's, one field
  along (`edit/draft.ts`: a commit that would change nothing sends
  nothing), so it fences a gesture that would produce no write rather
  than a write. Comment corrected; the behaviour is right and unchanged.

Skipped, with the reasons the reviews are owed:

- **extracting the `Said` line** into a shared component. It is now
  spelled four times (`UndoSaid`, the `•••` menu, the row editor, the
  picker) and two reviews flagged it — but the three older ones differ in
  geometry AND in whether their tone colours the border, so the shared
  part is two attributes and a text tone. Rewiring two surfaces this PR
  is not about, on the eve of review, buys less than it risks; the drift
  the new copy introduced (the missing `aria-live`) is fixed above, and
  the extraction is the follow-up.
- **`<Dynamic>` in `DateBadge`.** The efficiency lens wants two compiled
  branches under a `<Show>` (Dynamic costs two memos and a full prop
  spread per pill, and the condition never flips per call site); the
  altitude lens wants exactly what is there (one element, so one testid
  and one set of `data-` facts across four surfaces). Both are right; the
  cost is a few memos on the dated rows of one page and the benefit is
  the promise the badge exists to keep, so it stays.
- **a page-level "which row is picking" cell**, which two reviews reached
  from opposite directions — one per-row signal saved, and one open
  picker at a time instead of N. It is the shape `editing.tsx`'s caret
  already has, and the honest version of it widens that caret rather than
  adding a parallel one, which is a refactor of the editor rather than a
  cleanup of this diff. Recorded for the reviewer; the per-row local
  state has direct precedent in `createNoteExpand()` two lines above it.
- **`isDay`/`hasTime` in `@olai/format`.** `noticeOf` asks "does this
  value carry a time" as `stored !== dayOf(stored)` — derived from the
  format's own function rather than a second reading of its rule, so the
  vocabulary is not duplicated, only used.

Green: typecheck, 1149 unit tests, and the four e2e features these
touched (setting_a_date, menu_verbs, undo, keyboard_editing) — 83
scenarios.
